Reverse swipe gestures to match osx
In this direction swiping mimics dragging workspaces so they slide
through the screen. This has become a habit because on OSX there is
a matching animation.

In the other direction the waybar workspace tray underline appears to
drag correctly. This means it is now confusing if looking there while
initiating a gesture.