Vim-startify to reload sessions/files at startup
Adds vim startify plugin to quickly reload files and sessions.

Learnt that :mksession can be used to save which files/tabs/splits
are open in vim to restor later. These can be retrieved with :source
or can be started by passing vim the "-S" flag.

Startify's :SSave command can be used to tell :mksession to use the
default session directory instead of the current working directory.